requirements.txt
Flask==2.2.2
gunicorn==20.1.0
opencv-python==4.6.0.66
Pillow==9.2.0
pytesseract==0.3.10
